# packages/api-client/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/me/timeline": {
      "get": {
        "operationId": "TimelineController_myTimeline",
        "parameters": [
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Timeline page envelope"
          }
        },
        "summary": "Read the authenticated data subject's own timeline",
        "tags": [
          "timeline"
        ]
      }
    },
    "/api/patients/{patientRef}/timeline": {
      "get": {
        "operationId": "TimelineController_patientTimeline",
        "parameters": [
          {
            "name": "patientRef",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Timeline page envelope"
          }
        },
        "summary": "Read a patient's timeline (self or delegated)",
        "tags": [
          "timeline"
        ]
      }
    },
...
    "/api/appointments/{appointmentId}": {
      "get": {
        "operationId": "AppointmentsController_read",
        "parameters": [
          {
            "name": "appointmentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Appointment envelope"
          }
        },
        "summary": "Read one of the caller's own appointments",
        "tags": [
          "appointments"
        ]
      }
    },
    "/api/appointments/{appointmentId}/reschedule": {
      "post": {
        "operationId": "AppointmentsController_reschedule",
        "parameters": [
          {
            "name": "appointmentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reschedule result envelope"
          }
        },
        "summary": "Reschedule one of the caller's own appointments",
        "tags": [
          "appointments"
        ]
      }
    },
    "/api/appointments/{appointmentId}/cancel": {
      "post": {
        "operationId": "AppointmentsController_cancel",
        "parameters": [
          {
            "name": "appointmentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cancel result envelope"
          }
        },
        "summary": "Cancel one of the caller's own appointments",
        "tags": [
          "appointments"
        ]
      }
    }
  },
  "info": {
    "title": "NelyoHealth API Skeleton",
    "description": "Phase 2 API skeleton OpenAPI contract",
    "version": "0.0.0",
    "contact": {}
  },
...
  "servers": [],
...
}